+++ common/src/main/java/smartin/miapi/client/gui/crafting/statdisplay/AttributeSingleDisplay.java
@@ -4,6 +4,7 @@ import com.google.common.collect.Multimap;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.entity.attribute.AttributeContainer;
 import net.minecraft.entity.attribute.EntityAttribute;
 import net.minecraft.entity.attribute.EntityAttributeModifier;
 import net.minecraft.item.ItemStack;
@@ -16,11 +17,15 @@ import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
 import java.util.*;
 
+import static net.minecraft.entity.attribute.EntityAttributeModifier.Operation.MULTIPLY_BASE;
+import static net.minecraft.entity.attribute.EntityAttributeModifier.Operation.MULTIPLY_TOTAL;
+
 @Environment(EnvType.CLIENT)
 public class AttributeSingleDisplay extends SingleStatDisplayDouble {
     public static Set<EntityAttribute> attributesWithDisplay = new HashSet<>();
     public static Map<EquipmentSlot, Multimap<EntityAttribute, EntityAttributeModifier>> oldItemCache = new WeakHashMap<>();
     public static Map<EquipmentSlot, Multimap<EntityAttribute, EntityAttributeModifier>> compareItemCache = new WeakHashMap<>();
+    public EntityAttributeModifier.Operation operation = EntityAttributeModifier.Operation.ADDITION;
     final EntityAttribute attribute;
     final EquipmentSlot slot;
     double defaultValue;
@@ -65,7 +70,32 @@ public class AttributeSingleDisplay extends SingleStatDisplayDouble {
             for (EquipmentSlot equipmentSlot : EquipmentSlot.values()) {
                 Multimap<EntityAttribute, EntityAttributeModifier> currentSlot = attributeCache.get(equipmentSlot);
                 if (attributeCache.get(equipmentSlot).containsKey(attribute)) {
-                    return AttributeProperty.getActualValue(currentSlot, attribute, defaultValue);
+                    switch (operation) {
+                        case ADDITION -> {
+                            value = defaultValue;
+                            for (EntityAttributeModifier modifier : currentSlot.get(attribute).stream().filter(a -> a.getOperation().equals(operation)).toList()) {
+                                value += modifier.getValue();
+                            }
+                            return value;
+                        }
+                        case MULTIPLY_BASE -> {
+                            value = 0.0;
+                            for (EntityAttributeModifier modifier : currentSlot.get(attribute).stream().filter(a -> a.getOperation().equals(MULTIPLY_BASE)).toList()) {
+                                value += modifier.getValue();
+                            }
+                            for (EntityAttributeModifier modifier : currentSlot.get(attribute).stream().filter(a -> a.getOperation().equals(MULTIPLY_TOTAL)).toList()) {
+                                value = (value + 1) * (modifier.getValue() + 1) - 1;
+                            }
+                            return value * 100;
+                        }
+                        case MULTIPLY_TOTAL -> {
+                            value = 1.0;
+                            for (EntityAttributeModifier modifier : currentSlot.get(attribute).stream().filter(a -> a.getOperation().equals(operation)).toList()) {
+                                value = value * modifier.getValue();
+                            }
+                            return value * 100;
+                        }
+                    }
                 }
             }
             return defaultValue;
@@ -91,11 +121,9 @@ public class AttributeSingleDisplay extends SingleStatDisplayDouble {
             for (EquipmentSlot equipmentSlot : EquipmentSlot.values()) {
                 Multimap<EntityAttribute, EntityAttributeModifier> slotMap = attributeCache.get(equipmentSlot);
                 if (slotMap != null) {
-                    Collection<EntityAttributeModifier> attrCollection = attributeCache.get(equipmentSlot).get(attribute);
-                    if (attrCollection != null && !attrCollection.isEmpty()) {
-                        if (AttributeProperty.getActualValue(attrCollection, attribute.getDefaultValue()) != attribute.getDefaultValue()) {
-                            return true;
-                        }
+                    Collection<EntityAttributeModifier> attrCollection = attributeCache.get(equipmentSlot).get(attribute).stream().filter(attribute -> attribute.getOperation().equals(operation)).toList();
+                    if (hasValue(attrCollection)) {
+                        return true;
                     }
                 }
             }
@@ -103,12 +131,20 @@ public class AttributeSingleDisplay extends SingleStatDisplayDouble {
             Multimap<EntityAttribute, EntityAttributeModifier> slotMap = attributeCache.get(slot);
             if (slotMap != null) {
                 Collection<EntityAttributeModifier> attrCollection = attributeCache.get(slot).get(attribute);
-                if (attrCollection != null && !attrCollection.isEmpty()) {
-                    if (AttributeProperty.getActualValue(attrCollection, attribute.getDefaultValue()) != attribute.getDefaultValue()) {
-                        return true;
-                    }
-                }
+                return hasValue(attrCollection);
             }
+        }
+        return false;
+    }
+
+    public boolean hasValue(Collection<EntityAttributeModifier> list) {
+        list = list.stream().filter(attribute -> attribute.getOperation().equals(operation)).toList();
+        if (operation.equals(EntityAttributeModifier.Operation.ADDITION)) {
+            if (AttributeProperty.getActualValue(list, attribute.getDefaultValue()) != attribute.getDefaultValue()) {
+                return true;
+            }
+        } else {
+            return !list.isEmpty();
         }
         return false;
     }
@@ -210,7 +246,8 @@ public class AttributeSingleDisplay extends SingleStatDisplayDouble {
             double getValue(ItemStack itemStack);
         }
 
-        public AttributeSingleDisplay build() {
+        public AttributeSingleDisplay[] build() {
+            AttributeSingleDisplay[] displays = new AttributeSingleDisplay[2];
             // Validate the required fields
             if (name == null) {
                 throw new IllegalStateException("Name is required");
@@ -237,7 +274,52 @@ public class AttributeSingleDisplay extends SingleStatDisplayDouble {
                     }
                 };
             }
-            return display;
+            display.operation = EntityAttributeModifier.Operation.ADDITION;
+            displays[0] = display;
+
+            AttributeSingleDisplay displayMulBase = new AttributeSingleDisplay(attribute, slot, name, hoverDescription, defaultValue, modifierFormat);
+            displayMulBase.minValue = 0;
+            displayMulBase.maxValue = 100;
+            displayMulBase.setInverse(inverse);
+            if (valueGetter != null) {
+                displayMulBase.valueReader = new StatReaderHelper() {
+                    @Override
+                    public double getValue(ItemStack itemStack) {
+                        return valueGetter.getValue(itemStack);
+                    }
+
+                    @Override
+                    public boolean hasValue(ItemStack itemStack) {
+                        return displayMulBase.hasAttribute(itemStack);
+                    }
+                };
+            }
+            displayMulBase.operation = MULTIPLY_BASE;
+            displayMulBase.postfix = Text.literal("%");
+            displays[1] = displayMulBase;
+
+            AttributeSingleDisplay displayMulTotal = new AttributeSingleDisplay(attribute, slot, name, hoverDescription, defaultValue, modifierFormat);
+            displayMulTotal.minValue = 0;
+            displayMulTotal.maxValue = 100;
+            displayMulTotal.setInverse(inverse);
+            if (valueGetter != null) {
+                displayMulTotal.valueReader = new StatReaderHelper() {
+                    @Override
+                    public double getValue(ItemStack itemStack) {
+                        return valueGetter.getValue(itemStack);
+                    }
+
+                    @Override
+                    public boolean hasValue(ItemStack itemStack) {
+                        return displayMulTotal.hasAttribute(itemStack);
+                    }
+                };
+            }
+            displayMulTotal.operation = EntityAttributeModifier.Operation.MULTIPLY_TOTAL;
+            displayMulTotal.postfix = Text.literal("%");
+            //displays[2] = displayMulTotal;
+
+            return displays;
         }
     }
 }